Reorganize ZSS ACMD
