+++ src/fighter/szerosuit/acmd/mod.rs
@@ -1,0 +1,7 @@
+mod ground;
+mod aerials;
+
+pub fn install() {
+    ground::install();
+    aerials::install();
+}
